feat(cardswhitelist): Added a card-whitelist variable
Move the cards array to a angular.value so they can be overridden.

Also added in a cardWhitelist option that takes in an array of card type strings.

For some reason the cardType attribute ignores this but the types are still validated correctly so if a card type is not in the whitelist then the input is invalid.